+++ src/main/java/game/Player.java
@@ -1,0 +1,17 @@
+package game;
+
+import game.util.Observer;
+
+import java.util.Observable;
+
+public class Player implements Observer {
+    @Override
+    public void update(Observable o, Object arg) {
+
+    }
+
+    @Override
+    public void update() {
+
+    }
+}