+++ client/src/main/java/typeracer/client/ViewController.java
@@ -1,0 +1,8 @@
+package typeracer.client;
+
+/** Manages the user interface components and interactions. */
+public class ViewController {
+
+  /** The default constructor of this class. */
+  public ViewController() {}
+}