+++ client/src/main/java/typeracer/client/view/GameResultsUi.java
@@ -1,0 +1,8 @@
+package typeracer.client.view;
+
+/** Handles the display of game results. */
+public class GameResultsUi {
+
+  /** The default constructor of this class. */
+  public GameResultsUi() {}
+}