+++ client/src/main/java/typeracer/client/view/GameUi.java
@@ -1,0 +1,8 @@
+package typeracer.client.view;
+
+/** Handles the display of game. */
+public class GameUi {
+
+  /** The default constructor of this class. */
+  public GameUi() {}
+}