+++ client/src/main/java/typeracer/client/view/InitialPromptUi.java
@@ -1,0 +1,8 @@
+package typeracer.client.view;
+
+/** Handles the display of initial prompt. */
+public class InitialPromptUi {
+
+  /** The default constructor of this class. */
+  public InitialPromptUi() {}
+}