+++ client/src/main/java/typeracer/client/view/MainMenuUi.java
@@ -1,0 +1,8 @@
+package typeracer.client.view;
+
+/** Handles the display of main menu. */
+public class MainMenuUi {
+
+  /** The default constructor of this class. */
+  public MainMenuUi() {}
+}