+++ client/src/main/java/typeracer/client/view/ProfileSettingsUi.java
@@ -1,0 +1,8 @@
+package typeracer.client.view;
+
+/** Handles the display of profile setting. */
+public class ProfileSettingsUi {
+
+  /** The default constructor of this class. */
+  public ProfileSettingsUi() {}
+}