+++ game/src/main/java/typeracer/game/Player.java
@@ -1,0 +1,17 @@
+package typeracer.game;
+
+import typeracer.game.observable.Observable;
+import typeracer.game.observable.Observer;
+
+/** Represents a player of the game. */
+public class Player implements Observer {
+
+  /** The default constructor of this class. */
+  public Player() {}
+
+  @Override
+  public void update(Observable o, Object arg) {}
+
+  @Override
+  public void update() {}
+}