+++ game/src/main/java/typeracer/game/PlayerState.java
@@ -1,0 +1,8 @@
+package typeracer.game;
+
+/** Manages the state of a single player. */
+public class PlayerState {
+
+  /** The default constructor of this class. */
+  public PlayerState() {}
+}