+++ game/src/main/java/typeracer/game/TextSource.java
@@ -1,0 +1,8 @@
+package typeracer.game;
+
+/** Provides a text to use as prompt for the game. */
+public class TextSource {
+
+  /** The default constructor of this class. */
+  public TextSource() {}
+}