+++ game/src/main/java/typeracer/game/observable/Observer.java
@@ -1,0 +1,16 @@
+package typeracer.game.observable;
+
+/** Interface for objects that receive updates from an Observable. */
+public interface Observer {
+
+  /**
+   * Called when an Observable is updated.
+   *
+   * @param o the observable object
+   * @param arg an argument passed to the update method
+   */
+  void update(Observable o, Object arg);
+
+  /** Updates something. */
+  void update();
+}