+++ server/src/main/java/typeracer/server/GameController.java
@@ -1,0 +1,8 @@
+package typeracer.server;
+
+/** Controls the game logic and interactions. */
+public class GameController {
+
+  /** The default constructor of this class. */
+  public GameController() {}
+}