+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Assignment of tasks for each student (#6).
 - UML diagram.
 - Information about the GUI (#2).
+- Packages and the relevant classes based on MVC pattern
+- Observable and Observer in util package
+- build.gradle configuration
 
 ### Changed