+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Assignment of tasks for each student (#6).
 - UML diagram.
 - Information about the GUI (#2).
+- Information about the game logic (#11).
 
 ### Changed