+++ CHANGELOG.md
@@ -11,11 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - JSON protocol for Server-Client communication with sequence diagram (#3, #4).
-- Create README.md.
+- README.md.
 - Assignment of tasks for each student (#6).
 - UML diagram.
 - Information about the GUI (#2).
 - Information about the game logic (#11).
+- Packages and the relevant classes based on MVC pattern.
+- Observable and Observer in util package.
+- build.gradle configuration.
+- UI Classes.
+- Javadoc.
 
 ### Changed